+++ test/test_rackspace.py
@@ -55,8 +55,9 @@ class RackspaceTests(unittest.TestCase):
 
     def test_list_monitoring_zones(self):
         result = list(self.driver.list_monitoring_zones())
-        self.assertEqual(len(result), 1)
+        self.assertEqual(len(result), 2)
         self.assertEqual(result[0].id, 'mzxJ4L2IU')
+        self.assertEqual(result[1].id, 'pzdhaalpC2')
 
     def test_list_entities(self):
         result = list(self.driver.list_entities())
@@ -310,6 +311,11 @@ class RackspaceTests(unittest.TestCase):
     def test_views_metric_list(self):
         metric_list = self.driver.ex_views_metric_list()
         self.assertTrue(len(metric_list) > 0)
+
+    def test_delete_monitoring_zone_success(self):
+        monitoring_zone = self.driver.list_monitoring_zones()[1]
+        result = self.driver.delete_monitoring_zone(monitoring_zone=monitoring_zone)
+        self.assertTrue(result)
 
     def test_list_agent_tokens(self):
         tokens = self.driver.list_agent_tokens()